[core] Change dictionary module and refactor dictionary-std
- rename DictionaryEntry to pertain more about data element
- update stub dictionary module
- [dictionary-builder] adjust generated Rust code output accordingly
- [dictionary-std] move impl to data_element module
- [dictionary-std] only expose StandardDictionaryRegistry and StandardDataDictionary
  in crate root
- [dictionary-std] update tags module according to changes
